+++ app/themes/salon/booking-edit-modal.html.php
@@ -6,14 +6,17 @@
       <a class="modal-close" onclick="F1.Modal.close(this,event)">Close X</a>
     </header>
     <div class="modal-body">
-      <input type="hidden" name="id" value="">
+      <p class="field hidden">
+        <label>ID</label>
+        <input class="input" type="text" name="id" value="">
+      </p>
       <p class="field required" data-type="Select">
         <label>Client</label>
         <select 
           name="client_id" 
           data-locale="en" 
           data-search="true" 
-          data-value="0691234321"
+          data-value=""
           data-placeholder="- Select a client -">
           <?php foreach( $cal->clients as $c ): $lbl = "{$c->name}<small> - {$c->cell}</small>"; ?>
           <option value="<?=$c->id?>" title="<?=$lbl?>"><?="{$c->name} - {$c->cell}"?></option>
@@ -26,7 +29,7 @@
           name="treatment_id" 
           data-locale="en" 
           data-search="true" 
-          data-value="63"
+          data-value=""
           data-placeholder="- Select a treatment -">
           <?php foreach( $cal->treatments as $treatment ): ?>
           <option value="<?=$treatment->id?>"><?=$treatment->short_desc?></option>
@@ -41,7 +44,7 @@
               name="therapist_id" 
               data-locale="en" 
               data-search="true" 
-              data-value="2"
+              data-value=""
               data-placeholder="- Select a therapist -">
               <?php foreach( $cal->therapists as $t ): $lbl = "{$t->name}<small> - {$t->cell}</small>"; ?>
               <option value="<?=$t->id?>" title="<?=$lbl?>"><?="{$t->name} - {$t->cell}"?></option>
@@ -57,7 +60,7 @@
               name="station_id" 
               data-locale="en"
               data-search="true"
-              data-value="2"
+              data-value=""
               data-placeholder="- Select a station -">
               <!-- <optgroup title="Main Group"> -->
               <?php foreach( $cal->stations as $s ): $lbl = "STATION {$s->no}<small> - {$s->name}</small>"; ?>
@@ -84,9 +87,9 @@
             <div class="row" style="gap:0.67rem;align-items: flex-start;">
               <div class="input-container">
                 <select class="input" name="start_hour">
-                  <option value="7">7</option>
-                  <option value="08" selected="">08</option>
-                  <option value="09" >09</option>
+                  <option value="07">07</option>
+                  <option value="08">08</option>
+                  <option value="09">09</option>
                   <option value="10">10</option>
                   <option value="11">11</option>
                   <option value="12">12</option>
@@ -102,10 +105,10 @@
               <label style="padding:3px 0">:</label>
               <div class="col input-container" style="flex:1">
                 <select class="input" name="start_min">
-                  <option value="00" selected="">00</option>
-                  <option value="15" >15</option>
-                  <option value="30" >30</option>
-                  <option value="45" >45</option>
+                  <option value="00">00</option>
+                  <option value="15">15</option>
+                  <option value="30">30</option>
+                  <option value="45">45</option>
                 </select>
               </div>
             </div>
@@ -113,7 +116,7 @@
           <p class="field required" data-type="Duration">
             <label>Duration</label>
             <span class="range-input">
-              <input class="input" name="duration" type="range" value="90" min="15" max="120" step="15" list="dur">
+              <input class="input" name="duration" type="range" value="0" min="0" max="120" step="15" list="dur">
               <label class="duration_disp">15min</label>
               <datalist id="dur">
               <option value="15"  label="15min"></option>